Index thread elimination and optional partition size on create/rebuild.

+++ Katzebase.PublicLibrary/Payloads/KbBaseActionResponse.cs
@@ -11,7 +11,7 @@ namespace Katzebase.PublicLibrary.Payloads
         public int RowCount { get; set; }
         public List<KbQueryResultMessage> Messages { get; set; } = new();
         public Dictionary<KbTransactionWarning, HashSet<string>> Warnings { get; set; } = new();
-        public int Duration { get; set; } = 0;
+        public double Duration { get; set; } = 0;
 
         public KbBaseActionResponse()
         {
